move registerCommand to client
